+++ widgetry/src/lib.rs
@@ -56,9 +56,10 @@ pub use crate::widgets::filler::Filler;
 pub use crate::widgets::image::{Image, ImageSource};
 pub use crate::widgets::just_draw::DrawWithTooltips;
 pub(crate) use crate::widgets::just_draw::{DeferDraw, JustDraw};
-pub use crate::widgets::line_plot::{LinePlot, PlotOptions, Series};
+pub use crate::widgets::line_plot::LinePlot;
 pub use crate::widgets::menu::Menu;
 pub use crate::widgets::persistent_split::PersistentSplit;
+pub use crate::widgets::plots::{PlotOptions, Series};
 pub use crate::widgets::scatter_plot::ScatterPlot;
 pub use crate::widgets::slider::Slider;
 pub use crate::widgets::spinner::Spinner;
